+++ easyfarm/src/main/java/easyfarm/controller/PlanController.java
@@ -7,10 +7,15 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class PlanController {
 
-	@GetMapping("/plan")
-	public String planMain(Model model) {
-		
-		
-		return "views/plan/planMain";
+	/* 월켈린더조회 */
+	@GetMapping("/plan/getSchedule")
+	public String getSchedule(Model model) {
+		return "views/plan/getSchedule";
+	}
+	
+	/* 계획등록 */
+	@GetMapping("/plan/addSpend")
+	public String addSpend() {
+		return "views/plan/addSpend";
 	}
 }
